2026/misc/ctfaas/standalone_instancer/requirements.txt
Flask>=3.0.0
Flask-Limiter>=3.5.0
gunicorn>=22.0.0
pycryptodome>=3.19.0
requests>=2.31.0
python-dotenv>=1.0.0
redis>=5.0.0
